agregamos codigo js para la funcion de añadir peliculas favoritas, ahora el usuario puede saber si la pelicula que añadio esta repetida

+++ script.js
@@ -2,7 +2,7 @@ $(document).ready(function () {
     var movies = [];
     var jsonUrl = "movies.json";
 
-
+    
     function loadMovies() {
         $.ajax({
             url: jsonUrl,
@@ -17,11 +17,10 @@ $(document).ready(function () {
         });
     }
 
-
+    
     function displayMovies(movies) {
         var moviesList = $("#searchResults");
         moviesList.empty();
-
 
         movies.forEach(function (movie) {
             var movieItem = $("<div>");
@@ -31,48 +30,83 @@ $(document).ready(function () {
             var movieYear = $("<p>").text("Año: " + movie.year);
             var movieGenre = $("<p>").text("Género: " + movie.genre);
 
-
             var addToFavoritesButton = $("<button>").text("Agregar a favoritos");
             addToFavoritesButton.click(function () {
                 addToFavorites(movie);
             });
 
-
             movieItem.append(movieImage, movieTitle, movieDirector, movieYear, movieGenre, addToFavoritesButton);
-
-
             moviesList.append(movieItem);
         });
     }
 
-
+    
     function addToFavorites(movie) {
         var favorites = JSON.parse(localStorage.getItem("favorites")) || [];
+    
+        
+        var isDuplicate = favorites.some(function(favorite) {
+            return favorite.title === movie.title;
+        });
+    
+        
+        if (isDuplicate) {
+            Swal.fire({
+                icon: 'warning',
+                title: '¡Atención!',
+                text: 'Esa película ya la has agregado',
+                confirmButtonColor: '#3085d6',
+                confirmButtonText: 'Entendido'
+            });
+            return; 
+        }
+    
+        
         favorites.push(movie);
         localStorage.setItem("favorites", JSON.stringify(favorites));
         displayFavorites();
     }
 
-
     function displayFavorites() {
         var favoritesList = $("#favoritesList");
         favoritesList.empty();
-
+    
         var favorites = JSON.parse(localStorage.getItem("favorites")) || [];
-
+    
         favorites.forEach(function (movie) {
             var favoriteItem = $("<div>");
             var movieTitle = $("<span>").text(movie.title);
+            var movieThumbnail = $("<img>").attr("src", movie.imagen).attr("alt", movie.title).css({"width": "55px", "height": "60px"});
             var removeButton = $("<button>").text("Eliminar de favoritos");
             removeButton.click(function () {
                 removeFromFavorites(movie);
             });
-            favoriteItem.append(movieTitle, removeButton);
+            favoriteItem.append(movieThumbnail, movieTitle, removeButton);
             favoritesList.append(favoriteItem);
         });
     }
+    
+    function displayFavorites() {
+        var favoritesList = $("#favoritesList");
+        favoritesList.empty();
+    
+        var favorites = JSON.parse(localStorage.getItem("favorites")) || [];
+    
+        favorites.forEach(function (movie) {
+            var favoriteItem = $("<div>");
+            var movieTitle = $("<span>").text(movie.title);
+            var movieThumbnail = $("<img>").attr("src", movie.imagen).attr("alt", movie.title).css({"width": "50px", "height": "auto"});
+            var removeButton = $("<button>").text("Eliminar de favoritos");
+            removeButton.click(function () {
+                removeFromFavorites(movie);
+            });
+            favoriteItem.append(movieThumbnail, movieTitle, removeButton);
+            favoritesList.append(favoriteItem);
+        });
+    }
+    
 
-
+    
     function removeFromFavorites(movie) {
         var favorites = JSON.parse(localStorage.getItem("favorites")) || [];
         var updatedFavorites = favorites.filter(function (favorite) {
@@ -82,9 +116,24 @@ $(document).ready(function () {
         displayFavorites();
     }
 
+    
+    $("#toggleFavoritesList").click(function () {
+        $("#favoritesList").toggleClass("show");
+    });
 
+    
+    $("#searchButton").click(function () {
+        var searchTerm = $("#searchInput").val().toLowerCase();
+        var filteredMovies = movies.filter(function (movie) {
+            return movie.title.toLowerCase().includes(searchTerm);
+        });
+        displayMovies(filteredMovies);
+    });
+
+    
     loadMovies();
 });
+
